python/qrcodes/requirements.txt
pillow
qrcode